+++ api-rest-java/src/main/java/samucaguimaraes/apirestjava/repository/PessoaRepository.java
@@ -3,6 +3,6 @@ package samucaguimaraes.apirestjava.repository;
 import org.springframework.data.repository.CrudRepository;
 import samucaguimaraes.apirestjava.model.PessoaModel;
 
-public interface PessoaRepository extends CrudRepository<PessoaModel, Integer> {
+public interface PessoaRepository extends CrudRepository<PessoaModel, Long> {
 
 }
